Minor package refactoring
