+++ packages/mds-stream-processor/connectors/kafka/index.ts
@@ -1,0 +1,2 @@
+export * from './kafka-stream-sink'
+export * from './kafka-stream-source'
